requirements.txt
Flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.0
google-cloud-texttospeech==2.14.1
replicate==0.20.0
requests==2.31.0
webdav3-py==1.1.3
Pillow==10.0.0
